requirement.txt
scipy
django
djangorestframework
dynamic-rest
numpy
biopython
